Wild Ride. Refactored to use hooks
I realized that if you have this:

```
<State>
  {(state) =>
    <Dispatch>
      {({ saveState }) =>
        <MyComponent onSave={saveState} {...state}/>
      }
    </Dispatch>
  }
</State>
```

If the state changes, this will not cause the dispatch component to
rerender, which means that the Child component will not re-render.

This is not just a problem with the redux render props... I think it is
a problem with the whole approach? If you use render-props and you nest
things so that an inner component uses a bunch of stuff from nested
parent render props... maybe the component will miss re-renders?

This is worth thinking about. Anyway, I have moved on to try using hooks.