[WIP] Convert UI to use sidebar nav
Initially I was using the bottom navigation, but I think sidebar nav
makes more sense for an application that will be primarily desktop.

This was tricky because of hooks. Hooks need to be in function
components, they can't be in classes. React Router has not caught up to
hooks. WithRouter is an HOC and is typed to accept only class
components. So since material-ui uses hooks, its components are
functions, and can't be withRoutered.

Something like that anyway I'm going to sleep.

 - refactor: add drawer for navigation and appbar