refactor: rename supply -> supply24Hr
Event Buildings have supply, but it should be supply24Hr. As part of
this refactor I also added a rehydrate action on event buildings so
that if a property name changes in the future, the default will be
written to that property.

It might be good to add a facility for "migrations" that move data from
old properties to new ones.